starter/sample_submission/requirements.txt
azure-ai-formrecognizer
pandas
opencv-python
azure-cognitiveservices-vision-face
azure-storage-blob
tqdm